refactor: split cli/core.py (1963 lines) into focused submodules (fixes #635)
Split the monolithic cli/core.py into 9 focused modules under
naturo/cli/core/:

| Module         | Lines | Responsibility                          |
|----------------|-------|-----------------------------------------|
| _common.py     |    87 | Shared helpers, platform checks         |
| _capture.py    |   293 | capture command                         |
| _see.py        |   531 | see command                             |
| _find.py       |   334 | find command + AI find                  |
| _list.py       |   192 | list command group (apps/windows/screens)|
| _menu.py       |   129 | menu-inspect command                    |
| _learn.py      |   256 | learn command                           |
| _highlight.py  |   155 | highlight command                       |
| _tools.py      |    21 | tools command (hidden)                  |

Backward compatibility preserved via __init__.py re-exports.
Updated 10 test files (67 mock patch paths) from
naturo.cli.core._ to naturo.cli.core._common._ for shared helpers.
