+++ tests/test_error_format.py
@@ -22,8 +22,8 @@ class TestErrorFormatConsistency:
 
     def test_see_nonexistent_app_error_format(self, runner):
         """see --app nonexistent should output 'Error:' to stderr."""
-        with patch("naturo.cli.core._platform_supports_gui", return_value=True):
-            with patch("naturo.cli.core._get_backend") as mock_be:
+        with patch("naturo.cli.core._common._platform_supports_gui", return_value=True):
+            with patch("naturo.cli.core._common._get_backend") as mock_be:
                 backend = MagicMock()
                 backend._resolve_hwnds.return_value = []
                 mock_be.return_value = backend
@@ -35,21 +35,21 @@ class TestErrorFormatConsistency:
 
     def test_see_platform_error_format(self, runner):
         """see on unsupported platform should output 'Error:' to stderr."""
-        with patch("naturo.cli.core._platform_supports_gui", return_value=False):
+        with patch("naturo.cli.core._common._platform_supports_gui", return_value=False):
             result = runner.invoke(main, ["see"])
             assert result.exit_code != 0
             assert "Error:" in result.output
 
     def test_capture_platform_error_format(self, runner):
         """capture on unsupported platform should output 'Error:' to stderr."""
-        with patch("naturo.cli.core._platform_supports_gui", return_value=False):
+        with patch("naturo.cli.core._common._platform_supports_gui", return_value=False):
             result = runner.invoke(main, ["capture"])
             assert result.exit_code != 0
             assert "Error:" in result.output
 
     def test_list_windows_platform_error_format(self, runner):
         """list windows on unsupported platform should output 'Error:' to stderr."""
-        with patch("naturo.cli.core._platform_supports_gui", return_value=False):
+        with patch("naturo.cli.core._common._platform_supports_gui", return_value=False):
             result = runner.invoke(main, ["list", "windows"])
             assert result.exit_code != 0
             assert "Error:" in result.output
@@ -62,21 +62,21 @@ class TestErrorFormatConsistency:
 
     def test_click_no_target_error_format(self, runner):
         """click with no target should output 'Error:' to stderr."""
-        with patch("naturo.cli.core._platform_supports_gui", return_value=True):
+        with patch("naturo.cli.core._common._platform_supports_gui", return_value=True):
             result = runner.invoke(main, ["click"])
             assert result.exit_code != 0
             assert "Error:" in result.output
 
     def test_find_platform_error_format(self, runner):
         """find on unsupported platform should output 'Error:' to stderr."""
-        with patch("naturo.cli.core._platform_supports_gui", return_value=False):
+        with patch("naturo.cli.core._common._platform_supports_gui", return_value=False):
             result = runner.invoke(main, ["find", "button"])
             assert result.exit_code != 0
             assert "Error:" in result.output
 
     def test_menu_inspect_platform_error_format(self, runner):
         """menu-inspect on unsupported platform should output 'Error:' to stderr."""
-        with patch("naturo.cli.core._platform_supports_gui", return_value=False):
+        with patch("naturo.cli.core._common._platform_supports_gui", return_value=False):
             result = runner.invoke(main, ["menu-inspect"])
             assert result.exit_code != 0
             assert "Error:" in result.output
